requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
aiohttp==3.9.1
